Remove version management
